tasks/31_add-hybrid-search-capabilities/requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# BM25
rank-bm25>=0.2.2

# Semantic Search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Text Processing
nltk>=3.8.0
spacy>=3.5.0

# Query Expansion
wordnet

# Data handling
pandas>=1.3.0
pickle5>=0.0.11

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.65.0
requests>=2.28.0
